chore: move functions to base_funcs
